layout/generic/test/test_grid_track_sizing_algo_001.html

<!DOCTYPE HTML>
<!--
     Any copyright is dedicated to the Public Domain.
     http://creativecommons.org/publicdomain/zero/1.0/
-->
<html><head>
  <meta charset="utf-8">
  <title>CSS Grid Test: intrinsic track sizing with spanning items</title>
  <link rel="author" title="Mats Palmgren" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1425599">
  <link rel="help" href="https://drafts.csswg.org/css-grid/#algo-content">
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <style type="text/css">
body,html { color:black; background:white; font-family:monospace; font-size:16px; padding:0; margin:0; }
* { vertical-align: top; line-height: 1px; }

.grid {
  display: inline-grid;
  grid-template-rows: 1px;
  grid-auto-rows: 10px;
  border: 1px solid;
  place-content: start;
  place-items: start;
}

x {
  grid-row: 1;
  height: 1px;
  background: grey;
}

x:nth-child(1) { background: lime; width: 3px; grid-column: 1/span 2; }
x:nth-child(2) { background: silver; width: 6px; grid-column: 2/span 4; }
x:nth-child(3) { background: blue; width: 12px; grid-column: 2/span 2; }
  </style>
</head>
<body>

<script>
let min1 = [
  "0", "4px"
];
let min2 = [
  "auto", "min-content", "max-content"
];
let max1 = [
  "8px"
];
let max2 = [
  "auto", "min-content", "max-content"
];

var track = [];
min1.forEach((min) => { max2.forEach((max) => {
  track.push("minmax(" + min + "," + max + ")");
})});
min2.forEach((min) => { max2.forEach((max) => {
  track.push("minmax(" + min + "," + max + ")");
})});
min2.forEach((min) => { max1.forEach((max) => {
  track.push("minmax(" + min + "," + max + ")");
})});
min2.forEach((min) => { max2.forEach((max) => {
  track.push("minmax(" + min + "," + max + ")");
})});
track.push("fit-content(8px)");

var cols = [];
track.forEach((c1) =>  {
  cols.push(c1 + " " + c1 + " " + c1 + " " + c1 + " " + c1);
  track.forEach((c2) => {
    if (c1 != c2) {
      cols.push(c1 + " " + c2 + " " + c2 + " " + c2 + " " + c2);
      cols.push(c2 + " " + c1 + " " + c1 + " " + c2 + " " + c2);
    }
})});

document.body.style.display = 'none';
cols.forEach((col) => {
  let grid = document.createElement('div');
  grid.className = "grid";
  grid.style.gridTemplateColumns = col;
  grid.appendChild(document.createElement('x'))
  grid.appendChild(document.createElement('x'))
  grid.appendChild(document.createElement('x'))
  document.body.appendChild(grid);
});
document.body.style.display = '';
</script>

<script>
var actual = [];
[...document.querySelectorAll('.grid')].forEach(function(e) {
  let cs = window.getComputedStyle(e);
  actual.push([cs.gridTemplateColumns, cs.width]);
});

function dumpResult() {
  var s = "";
  actual.forEach(v => {
    s += '["' + v[0] + '","' + v[1] + '"],\n';
  });
  let pre = document.createElement("pre");
  pre.innerHTML = s;
  document.body.appendChild(pre);
}

let expected = [
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 8px 8px 8px 8px","32px"],
["8px 6px 6px 8px 8px","36px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 3px 3px","21px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 4px 4px","24px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 8px 8px 8px 8px","36px"],
["8px 6px 6px 8px 8px","36px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["4px 6px 6px 0px 0px","16px"],
["0px 6px 6px 0px 0px","12px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["8px 8px 8px 8px 8px","40px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 8px 8px 8px 8px","40px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 8px 8px 8px 8px","40px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 0px 0px","20px"],
["0px 8px 8px 0px 0px","16px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 4px 4px","28px"],
["4px 8px 8px 4px 4px","28px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 8px 8px 8px 8px","40px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["8px 6px 6px 0px 0px","20px"],
["1.5px 8px 8px 0px 0px","17.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["3px 6px 6px 0px 0px","15px"],
["0px 6px 6px 0px 0px","12px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["0px 6px 6px 4px 4px","20px"],
["4px 6px 6px 4px 4px","24px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 8px 8px 8px 8px","33.5px"],
["8px 6px 6px 8px 8px","36px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
["1.5px 6px 6px 0px 0px","13.5px"],
];

function test() {
  ok(expected.length > 0, "sanity check");
  for (i = 0; i < expected.length; ++i) {
    var msg = "";
    if (actual[i][0] != expected[i][0]) {
      msg = "'grid-template-columns' for grid index " + i + ":\n" + [...document.querySelectorAll('.grid')][i].outerHTML;
      is(actual[i][0], expected[i][0], msg);
    }
    if (actual[i][1] != expected[i][1]) {
      msg = "'width' for grid index " + i + ":\n" + [...document.querySelectorAll('.grid')][i].outerHTML;
      is(actual[i][1], expected[i][1], msg);
    }
    if (msg != "") {
      i = expected.length;  // stop after first failed grid by default
    }
  }
  SimpleTest.finish();
}

</script>

<script>
window.onload = function() { setTimeout(test, 0); };
SimpleTest.waitForExplicitFinish();
</script>

</body>
</html>
